fixup! Tests - Rename sql_test_base to db_setup
